+++ app/domain/services/optimize/indicator_type_optimizer.py
@@ -37,6 +37,8 @@ class IndicatorTypeOptimizationService:
             selected_indicators)
 
         settings = {
+            "broker": "Yahoo",
+          	"trading_period": "1d",
             "asset": "SPY",
             "start_date": "2010-01-01",
             "end_date": "2024-12-31",